Allocate polygon centroids to hexagons in a grid

Chooses a hexagon centroid for each polygon in the shape file, from a grid spanning the longitudes and latitudes in the expanded bounding box.

allocate(centroids, hex_grid, sf_id = names(centroids)[1], hex_size,
  hex_filter, focal_points = NULL, width, verbose)

Arguments

centroids
a data frame with centroids of non empty polygons

hex_grid
a data frame containing all possible hexagon points

sf_id
a string to indicate the column to identify individual polygons

hex_size
a float value in degrees for the diameter of the hexagons

hex_filter
amount of hexagons around centroid to consider

focal_points
a data frame of reference locations when allocating hexagons, capital cities of Australia are used in the example

width
a numeric indicating the angle used to filter the hexagon grid

verbose
a boolean to indicate whether to show polygon id

Value

a data frame of one allocation

Examples

# Create centroids set
centroids <- create_centroids(tas_lga, sf_id = "LGA_CODE16")

#> Warning: st_centroid assumes attributes are constant over geometries of x

#> Warning: st_centroid does not give correct centroids for longitude/latitude data

# Create hexagon location grid
data(capital_cities)
grid <- create_grid(centroids = centroids, hex_size = 0.2, buffer_dist = 1.2)
# Allocate polygon centroids to hexagon grid points
hex_allocated <- allocate(
  centroids = centroids,
  hex_grid = grid,
  hex_size = 0.2, # same size used in create_grid
  hex_filter = 10,
  focal_points = capital_cities,
  width = 30, verbose = TRUE
)

#> Allocating centroids, in order of distance to closest focal point.

#> 
#> Filter distance was expanded for  0 area(s).

# same column used in create_centroids
# create a set of hexagon points for plotting
fort_hex <- fortify_hexagon(data = hex_allocated, sf_id = "LGA_CODE16", hex_size = 0.2)
# plot the hexagons