Create a grid of evenly spaced points to allow hexagons to tesselate

This function takes the bounding box of a group of polygons, or a specific table of minimum and maximum longitudes and latitudes to create points for each polygon to be allocated to that will tesselate into hexagons.

create_grid(centroids, hex_size, buffer_dist, verbose = FALSE)

Arguments

centroids
data frame of centroids to be allocated

hex_size
a float value in degrees for the diameter of the hexagons

buffer_dist
distance to extend beyond the geometry provided

verbose
a boolean to indicate whether to show function progress

Value

grid

Examples

# Create a set of centroids for grid to overlay
centroids <- create_centroids(tas_lga, "LGA_CODE16")

#> Warning: st_centroid assumes attributes are constant over geometries of x

#> Warning: st_centroid does not give correct centroids for longitude/latitude data

# Create the grid
grid <- create_grid(centroids = centroids, hex_size = 0.2, buffer_dist = 1.2, verbose = FALSE)