Filter full set of grid points for those within range of original point

Takes only the closest available gridpoints as possible hexagon centroids to allocate polygons.

filter_grid_points(
  f_grid,
  f_centroid,
  focal_points = NULL,
  f_dist = filter_dist,
  angle_width = width,
  h_size = hex_size
)

Arguments

f_grid
complete grid of hexagon centroids

f_centroid
the longitude and latitude values for the current centroid

focal_points
a tibble of focal locations, an optional argument that allows allocation of polygons to hexagon centroids in ascending order of the distance to the closest focal point. It also filters the grid points to those within a 30 degree range of the angle from focal point to centroid. The default "capitals" uses the locations of the Australian capital cities as focal points.

f_dist
a distance in degrees, used as a boundary to filter the hexagon centroids considered for each polygon centroid to be allocated.

angle_width
a numeric used to filter the hexagon grid

h_size
a float value in degrees for the diameter of the hexagons

Value

a tibble of filtered grid points