Creates the points that define a hexagon polygon for plotting

Creates the points that define a hexagon polygon for plotting

fortify_hexagon(data, sf_id, hex_size)

Arguments

data
a data frame created by the allocate function

sf_id
a string to indicate the column to identify individual polygons

hex_size
a float value in degrees for the diameter of the hexagons

Value

a data frame of the seven points used to draw a hexagon

Examples

# Create centroids set
centroids <- create_centroids(tas_lga, "LGA_CODE16")

#> Warning: st_centroid assumes attributes are constant over geometries of x

#> Warning: st_centroid does not give correct centroids for longitude/latitude data

# Create hexagon location grid
grid <- create_grid(centroids = centroids, hex_size = 0.2, buffer_dist = 1.2)
# Allocate polygon centroids to hexagon grid points
allocated <- allocate(
  centroids = centroids,
  sf_id = "LGA_CODE16",
  hex_grid = grid,
  hex_size = 0.2, # same size used in create_grid
  hex_filter = 1,
  use_neighbours = tas_lga,
  width = 30,
  focal_points = capital_cities,
  verbose = TRUE
)

#> Allocating centroids, in order of distance to closest focal point.

#> although coordinates are longitude/latitude, st_intersects assumes that they are planar

#> 
#> Filter distance was expanded for  8 area(s).

# same column used in create_centroids
fortify_hexagon(data = allocated, sf_id = "LGA_CODE16", hex_size = 0.2)

#> # A tibble: 174 x 16
#>    LGA_CODE16 longitude latitude points focal_longitude focal_latitude
#>    <chr>          <dbl>    <dbl> <chr>            <dbl>          <dbl>
#>  1 60210           148.    -41.4 Hobart            147.          -42.9
#>  2 60210           148.    -41.4 Hobart            147.          -42.9
#>  3 60210           148.    -41.4 Hobart            147.          -42.9
#>  4 60210           148.    -41.4 Hobart            147.          -42.9
#>  5 60210           148.    -41.4 Hobart            147.          -42.9
#>  6 60210           148.    -41.4 Hobart            147.          -42.9
#>  7 60410           147.    -42.7 Hobart            147.          -42.9
#>  8 60410           147.    -42.7 Hobart            147.          -42.9
#>  9 60410           147.    -42.7 Hobart            147.          -42.9
#> 10 60410           147.    -42.7 Hobart            147.          -42.9
#> # … with 164 more rows, and 10 more variables: focal_dist <dbl>,
#> #   focal_angle <dbl>, rownumber <int>, hex_long <dbl>, hex_lat <dbl>,
#> #   hex_id <int>, long <dbl>, lat <dbl>, id <int>, poly_type <chr>