2019 Australian Federal election data: First preference votes for candidates (House of Representatives) in each electorate.

A dataset containing first preference vote counts, candidate names, and other results for the House of Representatives from the 2016 Australian federal election. The data were obtained from the Australian Electoral Commission, and downloaded from https://results.aec.gov.au/24310/Website/Downloads/HouseFirstPrefsByPartyDownload-24310.csv

fp19

Format

A data frame with the following variables:

StateAb

Abbreviation for state name

DivisionID

numeric identifier that links the electoral division with Census and other election datasets.

DivisionNm

Electoral division name

CandidateID

Candidate ID

Surname

Candidate surname

GivenNm

Candidate given name

BallotPosition

Candidate's position on the ballot

Elected

Whether the candidate was elected (Y/N)

HistoricElected

Whether the candidate is the incumbent member

PartyAb

Abbreviation for political party name

PartyNm

Political party name

OrdinaryVotes

Number of ordinary votes cast at the electorate for the candidate

AbsentVotes

Number of absentee votes

ProvisionalVotes

Number of provisional votes

PrePollVotes

Number of pre-poll votes

PostalVotes

Number of postal votes

TotalVotes

Number of total votes

Swing

% swing for or away from party