Read in the shape file as sf object

read_shape

read_shape(shp_path, simplify = TRUE, keep = 0.1)

Arguments

shp_path

character vector location of shape file, extension .shp

simplify

a boolean to decide whether to simplify the shape file using rmapshaper, keeping all shapes.

keep

ratio of points to keep

Value

an sf data frame, with a column of non null geometries

Examples

# \donttest{
# Example of how a shape file is read
if (interactive()) {
shape <- read_shape(shp_path = file.choose())
}
# }