Creates the points that define a hexagon polygon for plotting

Creates the points that define a hexagon polygon for plotting

fortify_hexagon(data, sf_id, hex_size)

Arguments

data

a data frame created by the allocate function

sf_id

a string to indicate the column to identify individual polygons

hex_size

a float value in degrees for the diameter of the hexagons

Value

a data frame of the seven points used to draw a hexagon

Examples

# same column is used in create_centroids
fortify_hexagon(data = tas_lga_hexctr, sf_id = "lga_code_2016", hex_size = 0.2)
#> # A tibble: 174 × 16
#>    lga_code_2016 longit…¹ latit…² points focal…³ focal…⁴ focal…⁵ focal…⁶ rownu…⁷
#>    <chr>            <dbl>   <dbl> <chr>    <dbl>   <dbl>   <dbl>   <dbl>   <int>
#>  1 60210             148.   -41.4 Hobart    147.   -42.9 170183.    21.2      17
#>  2 60210             148.   -41.4 Hobart    147.   -42.9 170183.    21.2      17
#>  3 60210             148.   -41.4 Hobart    147.   -42.9 170183.    21.2      17
#>  4 60210             148.   -41.4 Hobart    147.   -42.9 170183.    21.2      17
#>  5 60210             148.   -41.4 Hobart    147.   -42.9 170183.    21.2      17
#>  6 60210             148.   -41.4 Hobart    147.   -42.9 170183.    21.2      17
#>  7 60410             147.   -42.7 Hobart    147.   -42.9  14775.   -10.8       4
#>  8 60410             147.   -42.7 Hobart    147.   -42.9  14775.   -10.8       4
#>  9 60410             147.   -42.7 Hobart    147.   -42.9  14775.   -10.8       4
#> 10 60410             147.   -42.7 Hobart    147.   -42.9  14775.   -10.8       4
#> # … with 164 more rows, 7 more variables: hex_long <dbl>, hex_lat <dbl>,
#> #   hex_id <int>, long <dbl>, lat <dbl>, point_id <int>, poly_type <chr>, and
#> #   abbreviated variable names ¹​longitude, ²​latitude, ³​focal_longitude,
#> #   ⁴​focal_latitude, ⁵​focal_dist, ⁶​focal_angle, ⁷​rownumber